move EncryptRecipientDropdownAdapter to chips package
