+++ OpenKeychain/src/main/java/org/sufficientlysecure/keychain/ui/chips/EncryptRecipientChipsInput.java
@@ -9,7 +9,6 @@ import android.util.AttributeSet;
 import com.pchmn.materialchips.ChipsInput;
 import com.pchmn.materialchips.adapter.FilterableAdapter.FilterableItem;
 import org.sufficientlysecure.keychain.model.SubKey.UnifiedKeyInfo;
-import org.sufficientlysecure.keychain.ui.adapter.EncryptRecipientDropdownAdapter;
 import org.sufficientlysecure.keychain.ui.chips.EncryptRecipientChipsInput.EncryptRecipientChip;
